[CB-6] Split rendering items into seperate scripts
