+++ app/index.html
@@ -19,6 +19,6 @@
     </div>
 </div>
 
-<script src="renderer.js"></script>
+<script src="scripts/renderer.js"></script>
 </body>
 </html>
